Number addable
