+++ src/components/generic/editable-field-components/index.tsx
@@ -2,4 +2,4 @@ export { default as MetaformTextFieldComponent } from "./metaform-text-field-com
 export { default as MetaformHtmlComponent } from "./html-field-component";
 export { default as MetaformRadioFieldComponent } from "./metaform-radio-field-component";
 export { default as MetaformSubmitFieldComponent } from "./metaform-submit-field-component";
-export { default as MetaformNumberFieldComponent } from "./metaform-number-field-component";
+export { default as MetaformNumberFieldComponent } from "./number-field-component";
